Extract core classes into separate module
